move sort config into separate module
