+++ frontend/src/app/search/index.ts
@@ -1,5 +1,5 @@
 export * from './search.component';
 export * from './search-relevance.component';
 export * from './search-results.component';
-export * from './search-sorting.component';
+export * from './results-sort/search-sorting.component';
 export * from '../download/download.component';
