+++ frontend/src/app/search/results-sort/search-sorting.component.spec.ts
@@ -1,13 +1,13 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { mockCorpus } from '../../mock-data/corpus';
-import { commonTestBed } from '../common-test-bed';
-import { QueryModel } from '../models';
+import { mockCorpus } from '../../../mock-data/corpus';
+import { commonTestBed } from '../../common-test-bed';
+import { QueryModel } from '../../models';
 
 import { SearchSortingComponent } from './search-sorting.component';
-import { PageResults } from '../models/page-results';
-import { SimpleStore } from '../store/simple-store';
-import { SearchServiceMock } from '../../mock-data/search';
-import { SearchService } from '../services';
+import { PageResults } from '../../models/page-results';
+import { SimpleStore } from '../../store/simple-store';
+import { SearchServiceMock } from '../../../mock-data/search';
+import { SearchService } from '../../services';
 
 
 describe('Search Sorting Component', () => {
