+++ frontend/src/app/search/search.module.ts
@@ -6,11 +6,11 @@ import { SearchResultsComponent } from './search-results.component';
 import { SearchComponent } from './search.component';
 import { DocumentModule } from '../document/document.module';
 import { CorpusModule } from '../corpus-header/corpus.module';
-import { SearchSortingComponent } from './search-sorting.component';
 import { FilterModule } from '../filter/filter.module';
 import { DownloadModule } from '../download/download.module';
 import { QueryService, SearchService } from '../services';
 import { VisualizationModule } from '../visualization/visualization.module';
+import { ResultsSortModule } from './results-sort/results-sort.module';
 
 
 
@@ -24,7 +24,6 @@ import { VisualizationModule } from '../visualization/visualization.module';
         PaginationComponent,
         SearchComponent,
         SearchResultsComponent,
-        SearchSortingComponent,
     ],
     imports: [
         CorpusModule,
@@ -33,6 +32,7 @@ import { VisualizationModule } from '../visualization/visualization.module';
         FilterModule,
         SharedModule,
         VisualizationModule,
+        ResultsSortModule,
     ],
     exports: [
         SearchComponent,
